Kvalitetsindikator smerteendring

Indikatoren regner ut andelen pasienter som opplever nedgang i smertenivå fra første til siste tilsyn. Pasienten registrerer sterkeste og svakeste smerte i ro og bevegelse ved første og siste tilsyn. Vi bruker variablene `StSm12` og `StSm21` for sterkeste smerte og `SvSm12` og `SvSm21` for svakeste smerte. `12` viser til første tilsyn og `21` til siste tilsyn.

For å oppfylle kravet for nevner må pasienten ha svart ut smertenivå ved både første og siste tilsyn. For å oppfylle kravet for teller må det være en nedgang fra første til siste tilsyn.

ki_smerteendring(d, var)

Arguments

d

datasett som inneholder nødvendige variabler for å beregne smerteendring.

var

Tekststreng for å indikere om det er endring i `sterkeste` eller `svakeste` smerte vi skal hente ut. Aksepterte verdier er `sterkeste` eller `svakeste`.

Value

Returnerer inndata med to ekstra variabler `ki_krit_teller` og `ki_krit_nevner` som indikerer om kriterier for indikator er oppfylt.

Utdata kan videre mates inn i [rapwhale::aggreger_ki_prop()], gruppert etter ønske for å beregne andel for aktuelle undergrupper.

Examples

d = tibble::tibble(
StSm12 = c(0L, 1L, 2L, 5L, NA_integer_, NA_integer_),
StSm21 = c(0L, 0L, 0L, 0L, NA_integer_, 0L),
SvSm12 = c(0L, 0L, 0L, 0L, NA_integer_, 0L),
SvSm21 = c(0L, 0L, 0L, 0L, NA_integer_, 0L)
)

d_smerteendring_st = ki_smerteendring(d, "sterkeste")
d_smerteendrinig_sv = ki_smerteendring(d, "svakeste")